+++ src/app/styles/color.ts
@@ -1,0 +1,14 @@
+export const color = {
+  PRIMARY_BLUE: '#1D3152',
+  POINT_PINK: '#F14F4F',
+  WHITE: '#FFFFFF',
+  BLACK_100: '#121212',
+  BLACK_80: '#4D4D4DE',
+  GRAY_50: '#8a8888',
+  GRAY_30: '#D9D9D9',
+  BLUE_50: '#8E98A9',
+  BLUE_30: '#EDF0F5',
+  BLUE_10: '#FAFBFD',
+};
+
+export type ColorType = typeof color;
