+++ src/app/styles/device.ts
@@ -1,0 +1,6 @@
+export const device = {
+  MOBILE: '(max-width: 60rem)',
+  TABLET: '(min-width: 60rem) and (max-width: 122.3rem)',
+  DESKTOP: '(min-width: 122.4rem)',
+};
+export type DeviceType = typeof device;
