+++ src/app/styles/typo.ts
@@ -1,0 +1,52 @@
+export const typo = {
+  TITLE: `
+    font-family: "Pretendard";
+    font-weight: 600;
+    font-size: 2.4rem; 
+    line-height: 140%;
+  `,
+  SUBTITE_20: `
+    font-family: "Pretendard";
+    font-weight: 600;
+    font-size: 2.0rem;
+    line-height: 140%;
+  `,
+  SUBTITLE_14: `
+    font-family: "Pretendard";
+    font-weight: 600;
+    font-size: 1.4rem;
+    line-height: 140%;
+  `,
+  BODY_SEMIBOLD: `
+    font-family: "Pretendard";
+    font-weight: 600;
+    font-size: 1.6rem; 
+    line-height: 140%;
+  `,
+  BODY_MEDIUM: `
+    font-family: "Pretendard";
+    font-weight: 500;
+    font-size: 1.6rem; 
+    line-height: 140%;
+  `,
+  CAPTION_14: `
+    font-family: "Pretendard";
+    font-weight: 500;
+    font-size: 1.4rem; 
+    line-height: 140%;
+  `,
+  CAPTION_12: `
+    font-family: "Pretendard";
+    font-weight: 500;
+    font-size: 1.2rem;
+    line-height: 140%;
+  `,
+  CAPTION_10: `
+    font-family: "Pretendard";
+    font-weight: 500;
+    font-size: 1.0rem;
+    line-height: 140%;
+  `,
+};
+
+export type TypoType = typeof typo;
